Delered the old RRectangle, RMap and Coords helpers and renamed all of the BRectangle, BMap and BCoords classes to become RRectangle, RMap and Coords. RRectangle is now using BigInteger as the numerator in the rational values.

+++ src/MSS.Types/RMap/RSize.cs
@@ -1,14 +1,15 @@
 ﻿using MSS.Types.Base;
+using System.Numerics;
 
 namespace MSS.Types
 {
-	public class RSize : Size<long>
+	public class RSize : Size<BigInteger>
 	{
 		public int Exp { get; init; }
 
 		public RSize() : this(0, 0, 0) { }
 
-		public RSize(long width, long height, int exp) : base(width, height)
+		public RSize(BigInteger width, BigInteger height, int exp) : base(width, height)
 		{
 			Exp = exp;
 		}
